Moving card and its parts into its own package
